fix: move Minigolf variants to own folders so router registers them
